binance-kafka-ingest/requirements.txt
websocket-client
kafka-python
